+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Changed force spawn mechanic for Artemis and Selene, fixing issues with them.
+
 ## [0.3.0] - 2024-05-20
 
 ### Fixed